# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Iremeapp REST APIs",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/student/attendances/{school_id}/{class_id}": {
            "get": {
                "tags": [
                    "Get attendance list API"
                ],
                "operationId": "getAttendance",
                "parameters": [
                    {
                        "name": "school_id",
                        "in": "path",
                        "description": "School ID"
                    },
                    {
                        "name": "class_id",
                        "in": "path",
                        "description": "Class ID"
                    },
                    {
                        "name": "for",
                        "in": "query",
                        "description": "Attendance for"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort"
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort By"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attendence is successfully retrieved",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/student/attendances/import": {
            "post": {
                "tags": [
                    "Import Attendance API"
                ],
                "operationId": "importAttendanceStudents",
                "parameters": [
                    {
                        "name": "school_id",
                        "in": "path",
                        "description": "School ID"
                    },
                    {
                        "name": "classroom",
                        "in": "path",
                        "description": "Class ID"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "school_id",
                                    "class_id"
                                ],
                                "properties": {
                                    "school_id": {
                                        "type": "text"
                                    },
                                    "class_id": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Students are successfully imported",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/daily/attendance/{school_id}/{class_id}": {
            "get": {
                "tags": [
                    "Get attendance statistics API "
                ],
                "operationId": "getDailyAttendance",
                "parameters": [
                    {
                        "name": "school_id",
                        "in": "path",
                        "description": "School ID"
                    },
                    {
                        "name": "class_id",
                        "in": "path",
                        "description": "Class ID"
                    },
                    {
                        "name": "for",
                        "in": "query",
                        "description": "Attendance for"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Attendance statistics is successfully retrieved",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/student/attendances": {
            "post": {
                "tags": [
                    "Update Attendance API"
                ],
                "operationId": "updateAttendance",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "school_id",
                                    "class_id",
                                    "student_id"
                                ],
                                "properties": {
                                    "school_id": {
                                        "type": "text"
                                    },
                                    "class_id": {
                                        "type": "number"
                                    },
                                    "student_id": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Attendance is successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            },
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "name": "Authorization",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:8000//oauth/authorize",
                        "tokenUrl": "http://localhost:8000//oauth/token",
                        "refreshUrl": "http://localhost:8000//token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}